% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATCG.R
\name{SADEG.ACGT}
\alias{SADEG.ACGT}
\title{SADEG.ACGT}
\usage{
SADEG.ACGT(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
Table which first row represent number of each nucleotide and row 2 represent percentage of each nucleotide.
}
\description{
Number and frequencies of occurrence of each nucleotide.
}
\details{
The Number of occurrence of each nucleotide.(A, T, C, and G). Furtheremore, The frequencies of occurrence of each nucleotide. (A %, T %, C %, and G %).
}
\examples{
SADEG.ACGT("actagtcacgatcag")
}
\author{
Babak Khorsand
}

