% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAI.R
\name{SADEG.CAI}
\alias{SADEG.CAI}
\title{SADEG.CAI}
\usage{
SADEG.CAI(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
CAI
}
\description{
Codon Adaption Index
}
\details{
Geometric mean of the RSCU values presented as relative adaptiveness value (w), The CAI for a gene is then defined as the geometric mean of w values for codons in that gene excluding methionine, tryptophan, and stop codons.
}
\examples{
SADEG.CAI(Nucleotide_Sequence="atggctgctgcagcggccagtcacgatcagaggtaagttgtcgcagcatgt")
}
\author{
Babak Khorsand
}

