% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{SADEG.CodonUsage}
\alias{SADEG.CodonUsage}
\title{SADEG.CodonUsage}
\usage{
SADEG.CodonUsage(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
CodonUsage
}
\description{
Frequency of occurrence of each Amino acids codons.
}
\details{
Frequency of occurrence of each Amino acids codons.
}
\examples{
SADEG.CodonUsage(Nucleotide_Sequence="atggctgcagcggccagtcacgatcagaggtaagttgtc")
}
\author{
Babak Khorsand
}

