% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimer.R
\name{SADEG.Dimer}
\alias{SADEG.Dimer}
\title{SADEG.Dimer}
\usage{
SADEG.Dimer(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
Table which first row represent number of each Dimers and row 2 represent percentage of each Dimers.
}
\description{
Number and frequency of each Dimers
}
\details{
Count the number of each Dimers and calculate frequencies of them.
}
\examples{
SADEG.Dimer("actagtcacgatcag")
}
\author{
Babak Khorsand
}

