% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimerStackingEnergy.R
\name{SADEG.DimerStackingEnergy}
\alias{SADEG.DimerStackingEnergy}
\title{SADEG.DimerStackingEnergy}
\usage{
SADEG.DimerStackingEnergy(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
Sum of Stacking energy of each Dimers.
}
\description{
Sum of Stacking energy of each Dimers.
}
\details{
Count the number of each Dimers and multiply by related stacking energy.
}
\examples{
SADEG.DimerStackingEnergy("actagtcacgatcag")
}
\author{
Babak Khorsand
}

