% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC.R
\name{SADEG.GC}
\alias{SADEG.GC}
\title{SADEG.GC}
\usage{
SADEG.GC(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
GC content
}
\description{
GC content.
}
\details{
GC content : Sum of frequencies of G and C.
}
\examples{
SADEG.GC("actagtcacgatcag")
}
\author{
Babak Khorsand
}

