% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC1.R
\name{SADEG.GC1}
\alias{SADEG.GC1}
\title{SADEG.GC1}
\usage{
SADEG.GC1(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
GC1
}
\description{
GC1 content.
}
\details{
GC1 content : Sum of frequencies of G and C at the first position of each codon.
}
\examples{
SADEG.GC1(Nucleotide_Sequence="atggctgctgcagcggccagtcacgatcagaggtaagttgtcgcagcatgt")
}
\author{
Babak Khorsand
}

