% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC12.R
\name{SADEG.GC12}
\alias{SADEG.GC12}
\title{SADEG.GC12}
\usage{
SADEG.GC12(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
GC12
}
\description{
GC12 content.
}
\details{
GC12 content : Sum of frequencies of G and C at first and second position of each codon.
}
\examples{
SADEG.GC12(Nucleotide_Sequence="atggctgctgcagcggccagtcacgatcagaggtaagttgtcgcagcatgt")
}
\author{
Babak Khorsand
}

