% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC2.R
\name{SADEG.GC2}
\alias{SADEG.GC2}
\title{SADEG.GC2}
\usage{
SADEG.GC2(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
GC2
}
\description{
GC2 content.
}
\details{
GC2 content : Sum of frequencies of G and C at the second position of each codon.
}
\examples{
SADEG.GC2(Nucleotide_Sequence="atggctgctgcagcggccagtcacgatcagaggtaagttgtcgcagcatgt")
}
\author{
Babak Khorsand
}

