% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC3.R
\name{SADEG.GC3}
\alias{SADEG.GC3}
\title{SADEG.GC3}
\usage{
SADEG.GC3(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
GC3
}
\description{
GC3 content.
}
\details{
GC3 content : Sum of frequencies of G and C at the third position of each codon.
}
\examples{
SADEG.GC3(Nucleotide_Sequence="atggctgctgcagcggccagtcacgatcagaggtaagttgtcgcagcatgt")
}
\author{
Babak Khorsand
}

