% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReverseComplement.R
\name{SADEG.ReverseComplement}
\alias{SADEG.ReverseComplement}
\title{SADEG.ReverseComplement}
\usage{
SADEG.ReverseComplement(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
Reverse-Complement
}
\description{
Complement of the reversed sequence.
}
\details{
Complement of the reversed sequence. (atcagt -> Reverse: tgacta -> Reverse-Complement: actgat)
}
\examples{
SADEG.ReverseComplement("actagtcacgatcag")
}
\author{
Babak Khorsand
}

