% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCU.R
\name{SADEG.RSCU}
\alias{SADEG.RSCU}
\title{SADEG.RSCU}
\usage{
SADEG.RSCU(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
RSCU
}
\description{
Relative Synonymous Codon Usage
}
\details{
The Relative Synonymous Codon Usage (RSCU) is the number of times a codon appears in a gene divided by the number of expected occurrences under equal codon usage.
}
\examples{
SADEG.RSCU(Nucleotide_Sequence="atggctgcagcggccagtcacgatcagaggtaagttgtc")
}
\author{
Babak Khorsand
}

