% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackingEnergy.R
\name{SADEG.StackingEnergy}
\alias{SADEG.StackingEnergy}
\title{SADEG.StackingEnergy}
\usage{
SADEG.StackingEnergy(Nucleotide_Sequence)
}
\arguments{
\item{Nucleotide_Sequence}{Nucleotide Sequence}
}
\value{
StackingEnergy
}
\description{
Average free energy
}
\details{
Average free energy value was obtained via multiplying average energy of each dimer from each sequence in number of stacked bases. In order to designate and identify a product-favored process, minus sign was denoted before the average free energy (AFE) value. Lastly, all relevant stacked bases in each gene cumulatively defined the total stacking energy for respective sequences.
}
\examples{
SADEG.StackingEnergy("actagtcacgatcag")
}
\author{
Babak Khorsand
}

